% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnomonic3d.R
\name{do_gnomonic}
\alias{do_gnomonic}
\title{Illustrate gnomonic}
\usage{
do_gnomonic()
}
\value{
nothing, side effect a 3D plot
}
\description{
May need 'rgl::rglwidget()' if you don't have interactive graphics.
}
\examples{
#do_gnomonic()

}
