% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortho3d.r
\name{do_ortho}
\alias{do_ortho}
\title{Illustrate orthographic}
\usage{
do_ortho()
}
\value{
nothing, side effect a 3D plot
}
\description{
May need 'rgl::rglwidget()' if you don't have interactive graphics.
}
\examples{
do_ortho()
}
