% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereographic3d.R
\name{do_stereographic}
\alias{do_stereographic}
\title{Illustrate stereographic}
\usage{
do_stereographic()
}
\value{
nothing, side effect a 3D plot
}
\description{
May need 'rgl::rglwidget()' if you don't have interactive graphics.
}
\examples{
do_stereographic()
}
