% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000get-map.R
\name{get_map}
\alias{get_map}
\title{Get map coordinates}
\usage{
get_map(na.rm = TRUE)
}
\arguments{
\item{na.rm}{remove missing values or not (they indicate path start/ends)}
}
\value{
matrix of longlat coordinates of global national boundaries and coastlines (approximately)
}
\description{
Get map coordinates
}
\examples{
plot(get_map(), pch = ".")
lines(get_map(na.rm = FALSE))
}
