% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000llh2xyz.R
\name{llh2xyz}
\alias{llh2xyz}
\title{Projection longlat to geocentric (3D)}
\usage{
llh2xyz(lonlatheight, rad = 6370997, exag = 1)
}
\arguments{
\item{lonlatheight}{lonlat values (matrix, 2columns)}

\item{rad}{radius of sphere to use}

\item{exag}{exageration of the height to use}
}
\value{
matrix of 3 columns (geocentric x,y,z)
}
\description{
Projection longlat to geocentric (3D)
}
\examples{
llh2xyz(cbind(147, -42, 0))
}
